wsd: move Regex utilities to own home
