+++ Makefile.am
@@ -122,9 +122,10 @@ shared_sources = common/FileUtil.cpp \
                  common/Session.cpp \
                  common/Seccomp.cpp \
                  common/MobileApp.cpp \
-                 common/TraceEvent.cpp \
+                 common/RegexUtil.cpp \
                  common/SigUtil.cpp \
                  common/SpookyV2.cpp \
+                 common/TraceEvent.cpp \
                  common/Unit.cpp \
                  common/Uri.cpp \
                  common/Util.cpp \
@@ -385,6 +386,7 @@ shared_headers = common/Anonymizer.hpp \
                  common/LangUtil.hpp \
                  common/Log.hpp \
                  common/Protocol.hpp \
+                 common/RegexUtil.hpp \
                  common/StateEnum.hpp \
                  common/StringVector.hpp \
                  common/Seccomp.hpp \
