+++ android/lib/src/main/cpp/CMakeLists.txt.in
@@ -13,6 +13,7 @@ add_library(androidapp SHARED
             ../../../../../common/Log.cpp
             ../../../../../common/TraceEvent.cpp
             ../../../../../common/Protocol.cpp
+            ../../../../../common/RegexUtil.cpp
             ../../../../../common/Simd.cpp
             ../../../../../common/StringVector.cpp
             ../../../../../common/Session.cpp
