+++ common/CommandControl.cpp
@@ -14,6 +14,7 @@
 #include "CommandControl.hpp"
 
 #include <common/ConfigUtil.hpp>
+#include <common/RegexUtil.hpp>
 #include <common/Util.hpp>
 
 #include <string>
@@ -25,8 +26,8 @@ bool LockManager::_isLockedUser = false;
 bool LockManager::_isHostReadOnly = false;
 std::unordered_set<std::string> LockManager::LockedCommandList;
 std::string LockManager::LockedCommandListString;
-Util::RegexListMatcher LockManager::readOnlyWopiHosts;
-Util::RegexListMatcher LockManager::disabledCommandWopiHosts;
+RegexUtil::RegexListMatcher LockManager::readOnlyWopiHosts;
+RegexUtil::RegexListMatcher LockManager::disabledCommandWopiHosts;
 std::map<std::string, std::string> LockManager::unlockLinkMap;
 bool LockManager::lockHostEnabled = false;
 std::string LockManager::translationPath = std::string();
