+++ common/Util.cpp
@@ -773,103 +773,6 @@ namespace Util
         return info;
     }
 
-    bool matchRegex(const std::set<std::string>& set, const std::string& subject)
-    {
-        if (set.find(subject) != set.end())
-        {
-            return true;
-        }
-
-        // Not a perfect match, try regex.
-        for (const std::string& value : set)
-        {
-            try
-            {
-                // Not performance critical to warrant caching.
-                Poco::RegularExpression re(value, Poco::RegularExpression::RE_CASELESS);
-                Poco::RegularExpression::Match reMatch;
-
-                // Must be a full match.
-                if (re.match(subject, reMatch) && reMatch.offset == 0 &&
-                    reMatch.length == subject.size())
-                {
-                    return true;
-                }
-            }
-            catch (const std::exception& exc)
-            {
-                // Nothing to do; skip.
-            }
-        }
-
-        return false;
-    }
-
-    std::string getValue(const std::map<std::string, std::string>& map, const std::string& subject)
-    {
-        if (const auto& it = map.find(subject); it != map.end())
-        {
-            return it->second;
-        }
-
-        // Not a perfect match, try regex.
-        for (const auto& value : map)
-        {
-            try
-            {
-                // Not performance critical to warrant caching.
-                Poco::RegularExpression re(value.first, Poco::RegularExpression::RE_CASELESS);
-                Poco::RegularExpression::Match reMatch;
-
-                // Must be a full match.
-                if (re.match(subject, reMatch) && reMatch.offset == 0 &&
-                    reMatch.length == subject.size())
-                {
-                    return value.second;
-                }
-            }
-            catch (const std::exception& exc)
-            {
-                // Nothing to do; skip.
-            }
-        }
-
-        return std::string();
-    }
-
-    std::string getValue(const std::set<std::string>& set, const std::string& subject)
-    {
-        auto search = set.find(subject);
-        if (search != set.end())
-        {
-            return *search;
-        }
-
-        // Not a perfect match, try regex.
-        for (const auto& value : set)
-        {
-            try
-            {
-                // Not performance critical to warrant caching.
-                Poco::RegularExpression re(value, Poco::RegularExpression::RE_CASELESS);
-                Poco::RegularExpression::Match reMatch;
-
-                // Must be a full match.
-                if (re.match(subject, reMatch) && reMatch.offset == 0 &&
-                    reMatch.length == subject.size())
-                {
-                    return value;
-                }
-            }
-            catch (const std::exception& exc)
-            {
-                // Nothing to do; skip.
-            }
-        }
-
-        return std::string();
-    }
-
     void assertCorrectThread(std::thread::id owner, const char* fileName, int lineNo)
     {
         // uninitialized owner means detached and can be invoked by any thread.
