+++ common/Util.hpp
@@ -14,14 +14,12 @@
 #include <cassert>
 #include <cerrno>
 #include <chrono>
-#include <cinttypes>
 #include <cstddef>
 #include <cstdint>
 #include <cstring>
 #include <algorithm>
 #include <limits>
 #include <mutex>
-#include <set>
 #include <sstream>
 #include <string>
 #include <map>
@@ -944,71 +942,6 @@ int main(int argc, char**argv)
     /// Cleanup a filename replacing anything potentially problematic
     /// either for a URL or for a file path
     std::string cleanupFilename(const std::string &filename);
-
-    /// Return true if the subject matches in given set. It uses regex
-    /// Mainly used to match WOPI hosts patterns
-    bool matchRegex(const std::set<std::string>& set, const std::string& subject);
-
-    /// Return value from key:value pair if the subject matches in given map. It uses regex
-    /// Mainly used to match WOPI hosts patterns
-    std::string getValue(const std::map<std::string, std::string>& map, const std::string& subject);
-
-    std::string getValue(const std::set<std::string>& set, const std::string& subject);
-
-    /// Given one or more patterns to allow, and one or more to deny,
-    /// the match member will return true if, and only if, the subject
-    /// matches the allowed list, but not the deny.
-    /// By default, everything is denied.
-    class RegexListMatcher final
-    {
-    public:
-        RegexListMatcher() :
-            _allowByDefault(false)
-        {
-        }
-
-        explicit RegexListMatcher(const bool allowByDefault)
-            : _allowByDefault(allowByDefault)
-        {
-        }
-
-        void allow(const std::string& pattern) { _allowed.insert(pattern); }
-        void deny(const std::string& pattern)
-        {
-            _allowed.erase(pattern);
-            _denied.insert(pattern);
-        }
-
-        void clear()
-        {
-            _allowed.clear();
-            _denied.clear();
-        }
-
-        bool match(const std::string& subject) const
-        {
-            return (_allowByDefault ||
-                    Util::matchRegex(_allowed, subject)) &&
-                   !Util::matchRegex(_denied, subject);
-        }
-
-        /// Whether a match exist in either _allowed or _denied.
-        bool matchExist(const std::string& subject) const
-        {
-            return (Util::matchRegex(_allowed, subject) ||
-                    Util::matchRegex(_denied, subject));
-        }
-
-        bool empty() const
-        {
-            return _allowed.empty() && _denied.empty();
-        }
-
-    private:
-        std::set<std::string> _allowed;
-        std::set<std::string> _denied;
-        const bool _allowByDefault;
-    };
 
     /// Simple backtrace capture
     /// Use case, e.g. streaming up to 20 frames to log: `LOG_TRC( Util::Backtrace::get(20) );`
